Certification/backend/requirements.txt
# Core API framework
fastapi==0.115.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# LangChain and AI - use compatible versions
langchain==0.2.17
langchain-community==0.2.19
langchain-openai==0.1.25
langchain-core==0.2.43
langgraph==0.4.10

# Vector database
chromadb==0.5.5

# OpenAI
openai>=1.40.0,<2.0.0

# Evaluation
ragas>=0.1.0
pandas>=2.0.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.0.0,<3.0.0

# MCP (Model Context Protocol)
mcp==1.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0